feat: implement community search and preserve coaching pace precision

+++ src/main/java/com/neostride/server/coaching/repository/PlanDayInsertCommand.java
@@ -6,6 +6,6 @@ import java.time.LocalDate;
 public record PlanDayInsertCommand(
 		LocalDate planDate,
 		BigDecimal targetDistance,
-		Integer targetPace
+		BigDecimal targetPace
 ) {
 }
